feat: Publish zap receipt to relays from zap request.
Note, this will show some errors in dev as the ws://localhost:7000 relay can not be reached from within docker.

```
ERROR nostr_relay_pool::relay::internal: Impossible to get information document from ws://localhost:7000/: The relay information document is not accessible
``
